# backend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Backend
host: localhost:4000
...
basePath: /api
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.
  
  
     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.
  
  
     The following syntax must be used in the 'Authorization' header :
  
         Bearer xxxxxx.yyyyyyy.zzzzzz
consumes:
  - application/json
produces:
  - application/json
paths:
  /users:
    x-swagger-router-controller: users.controller
    post:
      tags:
        - Users
        - Login
      operationId: usersControllerPost
      description: "Create User"
      parameters:
        - name: authentication
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "201":
          description: "Created"
        "403":
          description: "Access Denied"
    get:
      tags:
        - Users
      operationId: usersControllerGet
      summary: "Get All The Users"
      description: "Get all the users"
      responses:
        200:
          description: "Success"
          schema:
            $ref: '#/definitions/User'
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
  /users/{idUser}:
    x-swagger-router-controller: users.controller
    parameters:
      - name: idUser
        in: path
        description: "ID of the User"
        required: true
        type: string
    get:
      tags:
        - Users
      operationId: usersControllerGetId
      summary: "Get an User"
      description: "Get only one User"
      responses:
        200:
          description: "Success"
          schema:
            $ref: '#/definitions/User'
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    
    put:
      tags:
        - Users
      operationId: usersControllerPut
      summary: "Update an User"
      description: "Update only one User"
      parameters:
        - name: User
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    
    delete:
      tags:
        - Users
      operationId: usersControllerDelete
      summary: "Delete an User"
      description: "Delete only one User"
      responses:
        200:
          description: "Success"
        403:
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
  /login:
    x-swagger-router-controller: users.controller
    post:
      tags:
        - Login
      operationId: loginPost
      description: "Login"
      parameters:
        - name: authentication
          in: body
          required: true
          schema:
            $ref: "#/definitions/Authentication"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/Login"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Authentication:
    type: object
    properties:
      username:
        type: string
        example: "username"
      password:
        type: string
        example: "password"
  User:
    type: object
    properties:
      dni:
        type: string
        example: "dni"
      firstname:
        type: string
        example: "firstname"
      lastname:
        type: string
        example: "lastname"
      birthdate:
        type: string
        example: "birthdate"
      username:
        type: string
        example: "username"
      password:
        type: string
        example: "password"
      email:
        type: string
        example: "example@exp.com"
      roles:
        type: array
        items:
          $ref: '#/definitions/Role'
  Role:
    type: object
    properties:
      name:
        type: string
        example: "resident"
      resources:
        $ref: '#/definitions/Resource'
      specialPermission:
        type: string
        enum: ["Full Access", "No Access", "Null"]
        example: "Full Access"
      description:
        type: string
        example: "This is the role of a resident"
  Resource:
    type: object
    properties:
      'question':
        $ref: '#/definitions/Permission'
      'answer':
        $ref: '#/definitions/Permission'
      'answer type':
        $ref: '#/definitions/Permission'
      'question group':
        $ref: '#/definitions/Permission'
      'service':
        $ref: '#/definitions/Permission'
      'institution':
        $ref: '#/definitions/Permission'
      'poll':
        $ref: '#/definitions/Permission'
      'user':
        $ref: '#/definitions/Permission'
      'role':
        $ref: '#/definitions/Permission'
      'procedure':
        $ref: '#/definitions/Permission'
      'patient':
        $ref: '#/definitions/Permission'
  Permission:
    type: object
    properties:
      view:
        type: boolean
        example: true
      create:
        type: boolean
        example: true
      delete:
        type: boolean
        example: false
      modify:
        type: boolean
        example: true
...
  Login:
    type: object
    properties:
      token:
        type: string
      client:
        type: string
  Error:
    type: object
    properties:
      message:
        type: string